refactor: Dockerfile and compose for development
we're moving to using the same Dockerimage for Celery workers, this drops the RUN and CMD
directives from the Dockerfile and instead uses the command directive in the compose
file to kickoff FastAPI or Celery, depending on the function.

the raltionale is that we will end up with one image per application and then use the command
and various other configuration to expose ports and or run apps, give that this is a container
this approach should be secure - I am yet to check this with the other engineers.