refactor: removes fastapi-jwt-auth (REFS#34)
fastapi-jwt-auth is a particularly nice implementaiton of jwt authentication
for fastapi, it's a middleware that does a lot heavy lifting for us, however
this has not been updated in over two years.

@devraj contacted the developer and did not get a response. there are a few
forks which are a stop gap solution, until such point we will have to maintain
a jwt imeplementation on our own.

dependabot has raised an issue with the version of pwjwt that fastapi-jwt-auth
uses and hence this decision is made, note that the issues on github have been
flagged as a security concern