refactor: establishes patterns for selectinload options (REFS #41)
asyncpg does not support joinload and requires queries to use the selectinload option
to construct queries which loads the relationship data.

it's likely that each model will have more than one getter, this pattern (documented)
allows you to construct the base query (which can be overridden) and then each
getter uses this inturn allowing you to maintain the selectinload code in one spot.