feat: begin work on users endpoints
intended to be used for an admin panel, we begin work on a set of endpoints to
manage users, this is also a demo of how CRUD operations can be performed with infinite
loading, keyword search, relationship loading and pagination.

the output of this will be ultimately used in a demo of a react-query based setup
on the client side to make use of these features.

upon completion these endpoints should be available to an admin style user