feat!: upate cryptography libraries REFS #58
BREAKING CHANGE: we depdend on passlib for cryptography, the package is unmaintained
and a third party package does not seem to have a roadmap. post research we are moving
to pyca maintained bcrypt which provides the same functionality.

we use hashpw and checkpw from the package, note these use byte strings not strings
which passlib would have provided a wrapper for